Consolidate ctl and zopectl
