+++ src/plone/recipe/zope2instance/ctl.py
@@ -28,15 +28,26 @@ available actions.
 """
 
 from pkg_resources import iter_entry_points
-from . import zopectl
-
+from ZConfig.loader import SchemaLoader
+from zdaemon.zdctl import ZDCmd, ZDCtlOptions
+from zdaemon.zdoptions import ZDOptions
+from Zope2.Startup.options import ConditionalSchemaParser
 import csv
 import os
 import os.path
+import pkg_resources
 import sys
+import xml.sax
+import zdaemon
 
+if sys.version_info > (3, ):
+    basestring = str
 
-if zopectl.WIN:
+WINDOWS = False
+if sys.platform[:3].lower() == "win":
+    WINDOWS = True
+
+if WINDOWS:
     import traceback
     from pkg_resources import resource_filename
     import pywintypes
@@ -51,15 +62,123 @@ if zopectl.WIN:
         'or you have not run the shell as Administrator.')
 
 try:
-    import ZServer
+    import ZServer  # noqa
     HAS_ZSERVER = True
 except ImportError:
     HAS_ZSERVER = False
 
+if HAS_ZSERVER:
+    from ZServer.Zope2.Startup.options import ZopeOptions
+else:
+    class ZopeOptions(ZDOptions):
+        schemadir = os.path.dirname(os.path.abspath(__file__))
+        schemafile = 'wsgischema.xml'
 
-class AdjustedZopeCmd(zopectl.ZopeCmd):
 
-    if zopectl.WIN:
+class ZopeCtlOptions(ZopeOptions, ZDCtlOptions):
+    # Zope controller options.
+    #
+    # After initialization, this should look very much like a
+    # zdaemon.zdctl.ZDCtlOptions instance.  Many of the attributes are
+    # initialized from different sources, however.
+
+    # Provide help message, without indentation.
+    __doc__ = __doc__
+
+    positional_args_allowed = True
+
+    # this indicates that no explicit program has been provided.
+    # the command line option can set this.
+    program = None
+
+    # this indicates that no explicit socket name has been provided.
+    # the command line option can set this.
+    sockname = None
+
+    # XXX Suppress using Zope's <eventlog> section to avoid using the
+    # same logging for zdctl as for the Zope appserver.  There still
+    # needs to be a way to set a logfile for zdctl.
+    logsectionname = None
+
+    def __init__(self):
+        ZopeOptions.__init__(self)
+        ZDCtlOptions.__init__(self)
+        self.add("interactive", None, "i", "interactive", flag=1)
+        self.add("default_to_interactive", "runner.default_to_interactive",
+                 default=1)
+
+    def realize(self, *args, **kw):
+        ZopeOptions.realize(self, *args, **kw)
+        # Additional checking of user option; set uid and gid
+        if self.user is not None:
+            import pwd
+            try:
+                uid = int(self.user)
+            except ValueError:
+                try:
+                    pwrec = pwd.getpwnam(self.user)
+                except KeyError:
+                    self.usage("username %r not found" % self.user)
+                uid = pwrec[2]
+            else:
+                try:
+                    pwrec = pwd.getpwuid(uid)
+                except KeyError:
+                    self.usage("uid %r not found" % self.user)
+            gid = pwrec[3]
+            self.uid = uid
+            self.gid = gid
+
+        config = self.configroot
+        self.directory = config.instancehome
+        self.clienthome = config.clienthome
+        if self.program:
+            if isinstance(self.program, basestring):
+                self.program = [self.program]
+        elif config.runner and config.runner.program:
+            self.program = config.runner.program
+        else:
+            self.program = [os.path.join(self.directory, "bin", "runzope")]
+        if self.sockname:
+            # set by command line option
+            pass
+        elif config.runner and config.runner.socket_name:
+            self.sockname = config.runner.socket_name
+        else:
+            self.sockname = os.path.join(self.clienthome, "zopectlsock")
+        self.python = os.environ.get('PYTHON', config.python) or sys.executable
+        self.zdrun = os.path.join(os.path.dirname(zdaemon.__file__),
+                                  "zdrun.py")
+
+        self.exitcodes = [0, 2]
+
+    def load_schema(self):
+        if self.schema is None:
+            # Load schema
+            if self.schemadir is None:
+                self.schemadir = os.path.dirname(__file__)
+            self.schemafile = os.path.join(self.schemadir, self.schemafile)
+            self._conditional_load()
+
+    def _conditional_load(self):
+        loader = SchemaLoader()
+        # loadURL
+        url = loader.normalizeURL(self.schemafile)
+        resource = loader.openResource(url)
+        try:
+            # load / parseResource without caching
+            parser = ConditionalSchemaParser(loader, resource.url)
+            xml.sax.parse(resource.file, parser)
+            self.schema = parser._schema
+        finally:
+            resource.close()
+
+
+class ZopeCmd(ZDCmd):
+
+    _exitstatus = 0
+
+    if WINDOWS:
 
         # printable representations of the Windows service states
         service_state_map = {
@@ -293,7 +412,7 @@ class AdjustedZopeCmd(zopectl.ZopeCmd):
                 print('ERROR: Zope is not installed as a Windows service.')
                 return
             elif status is not win32service.SERVICE_STOPPED:
-                print (
+                print(
                     'ERROR: Please stop the Windows service before '
                     'removing it.'
                 )
@@ -322,8 +441,8 @@ class AdjustedZopeCmd(zopectl.ZopeCmd):
             - Set "self.zd_up" to 1 or 0 (unclear what this is used for)
 
             """
-            zopectl.ZopeCmd.get_status(self)
-            # override value set by zopectl.ZopeCmd.get_status()
+            ZDCmd.get_status(self)
+            # override value set by ZDCmd.get_status()
             # (always -1 or 0)
             self.zd_pid = self._get_pid_from_pidfile()
 
@@ -360,8 +479,28 @@ class AdjustedZopeCmd(zopectl.ZopeCmd):
         def help_EOF(self):
             print('To quit, type CTRL+Z or use the quit command.')
 
-    # end of "if zopectl.WIN"
+    # end of "if WINDOWS"
     else:
+
+        def _get_override(self, opt, name, svalue=None, flag=0):
+            # Suppress the config file, and pass all configuration via the
+            # command line.  This avoids needing to specialize the zdrun
+            # script.
+            if name == "configfile":
+                return []
+            value = getattr(self.options, name)
+            if value is None:
+                return []
+            if flag:
+                if value:
+                    args = [opt]
+                else:
+                    args = []
+            else:
+                if svalue is None:
+                    svalue = str(value)
+                args = [opt, svalue]
+            return args
 
         def do_start(self, arg):
             self.get_status()
@@ -410,6 +549,69 @@ class AdjustedZopeCmd(zopectl.ZopeCmd):
                 cond,
                 'daemon process started, pid=%(zd_pid)d'
             )
+
+    def __getattr__(self, name):
+        """
+        Getter to check if an unknown command is implement by an entry point.
+        """
+        if not name.startswith("do_"):
+            raise AttributeError(name)
+        data = list(pkg_resources.iter_entry_points(
+            "zopectl.command", name=name[3:]))
+        if not data:
+            raise AttributeError(name)
+        if len(data) > 1:
+            sys.stderr.write(
+                "Warning: multiple entry points found for command")
+            return
+        func = data[0].load()
+        if not callable(func):
+            sys.stderr.write("Error: %s is not a callable method" % name)
+            return
+
+        return self.run_entrypoint(data[0])
+
+    def run_entrypoint(self, entry_point):
+        def go(arg):
+            # If the command line was something like
+            # """bin/instance run "one two" three"""
+            # cmd.parseline will have converted it so
+            # that arg == 'one two three'. This is going to
+            # foul up any quoted command with embedded spaces.
+            # So we have to return to self.options.args,
+            # which is a tuple of command line args,
+            # throwing away the "run" command at the beginning.
+            #
+            # Further complications: if self.options.args has come
+            # via subprocess, it may look like
+            # ['run "arg 1" "arg2"'] rather than ['run','arg 1','arg2'].
+            # If that's the case, we'll use csv to do the parsing
+            # so that we can split on spaces while respecting quotes.
+            tup = self.options.args
+            if len(tup) == 1:
+                tup = csv.reader(tup, delimiter=' ').next()
+
+            # Remove -c and add command name as sys.argv[0]
+            cmd = ['import sys',
+                   'sys.argv.pop()',
+                   'sys.argv.append(r\'%s\')' % entry_point.name
+                   ]
+            if len(tup) > 1:
+                argv = tup[1:]
+                for a in argv:
+                    cmd.append('sys.argv.append(r\'%s\')' % a)
+            cmd.extend([
+                'import pkg_resources',
+                'import Zope2',
+                'func=pkg_resources.EntryPoint.parse(\'%s\').load(False)'
+                % entry_point,
+                'app=Zope2.app()',
+                'func(app, sys.argv[1:])',
+            ])
+            cmdline = self.get_startup_cmd(
+                self.options.python, ' ; '.join(cmd))
+            self._exitstatus = os.system(cmdline)
+        return go
 
     def environment(self):
         configroot = self.options.configroot
@@ -484,7 +686,7 @@ class AdjustedZopeCmd(zopectl.ZopeCmd):
             cmdline += login_cmdline
 
         cmdline = cmdline + more + '\"'
-        if zopectl.WIN:
+        if WINDOWS:
             # entire command line must be quoted
             # as well as the components
             return '"%s"' % cmdline
@@ -541,7 +743,10 @@ class AdjustedZopeCmd(zopectl.ZopeCmd):
         self._exitstatus = os.system(cmdline)
 
     def help_run(self):
-        zopectl.ZopeCmd.help_run(self)
+        print("run <script> [args] -- run a Python script with the Zope ")
+        print("                       environment set up. The script can use ")
+        print("                       the name 'app' access the top-level ")
+        print("                       Zope object")
         self.help_startup_command()
 
     def do_console(self, arg):
@@ -564,12 +769,13 @@ console -- Run the program in the console.
             interactive_startup,
             pyflags='-i',
         )
-        print ('Starting debugger (the name "app" is bound to the top-level '
-               'Zope object)')
+        print('Starting debugger (the name "app" is bound to the top-level '
+              'Zope object)')
         os.system(cmdline)
 
     def help_debug(self):
-        zopectl.ZopeCmd.help_debug(self)
+        print("debug -- run the Zope debugger to inspect your database")
+        print("         manually using a Python interactive shell")
         self.help_startup_command()
 
     def do_foreground(self, arg, debug=True):
@@ -594,7 +800,7 @@ console -- Run the program in the console.
                 local_additions += ['debug-mode=on']
             program.extend(local_additions)
 
-        if zopectl.WIN:
+        if WINDOWS:
             # The outer quotes were causing
             # "WindowsError: [Error 87] The parameter is incorrect"
             # command = zopectl.quote_command(program)
@@ -602,7 +808,7 @@ console -- Run the program in the console.
         else:
             command = program
 
-        if debug or zopectl.WIN:
+        if debug or WINDOWS:
             try:
                 self._exitstatus = subprocess.call(command, env=env)
             except KeyboardInterrupt:
@@ -623,12 +829,33 @@ console -- Run the program in the console.
               "bin/test -s <my.package>")
         return
 
+    def do_adduser(self, arg):
+        try:
+            name, password = arg.split()
+        except Exception:
+            print("usage: adduser <name> <password>")
+            return
+        cmdline = self.get_startup_cmd(
+            self.options.python,
+            'import Zope2; '
+            'app = Zope2.app(); '
+            'result = app.acl_users._doAddUser('
+            '\'%s\', \'%s\', [\'Manager\'], []); '
+            'import transaction; '
+            'transaction.commit(); '
+            'print(\'Created user:\', result)'
+        ) % (name, password)
+        os.system(cmdline)
+
+    def help_adduser(self):
+        print("adduser <name> <password> -- add a Zope management user")
+
 
 def main(args=None):
     """Customized entry point for launching Zope without forking other processes
     """
 
-    options = zopectl.ZopeCtlOptions()
+    options = ZopeCtlOptions()
     options.add(name="no_request", short="R", long="no-request", flag=1)
     options.add(name="no_login", short="L", long="no-login", flag=1)
     options.add(name="object_path", short="O:", long="object-path=")
@@ -654,8 +881,7 @@ def main(args=None):
             options.python, options.interpreter, script, wsgi_ini
         ]
 
-    # We use our own ZopeCmd set, that is derived from the original one.
-    c = AdjustedZopeCmd(options)
+    c = ZopeCmd(options)
 
     # Mix in any additional commands supplied by other packages:
     for ep in iter_entry_points('plone.recipe.zope2instance.ctl'):
@@ -685,7 +911,7 @@ def main(args=None):
     # If no command was specified: enter interactive mode.
 
     try:
-        import readline
+        import readline  # noqa
     except ImportError:
         pass
     print('Program: {}'.format(' '.join(options.program)))
